+++ src/App.tsx
@@ -9,7 +9,7 @@ import ReviewList from './pages/ReviewList';
 import ReviewPage from './pages/ReviewPage';
 import SearchPage from './pages/SearchPage';
 import SigninPage from './pages/SigninPage';
-import SignUpPage from './pages/SignupPage';
+import SignUpPage from './pages/SignUpPage';
 import { reset } from './styles/Global';
 
 const router = createBrowserRouter([
